Initial GitHub Action implementation.
Implements initial action supporting a basic set of options to be passed
in.

The action consists of building documentation off any .md files found in
the repository and push them back as the gh-pages branch that can be
configured as github pages endpoint.

If the branch already exists it will create a branch with name:
 gh-pages-<workflow_run_id>-<workflow_run_count>

Optionally, if the branch name conflicts, it will increment a counter
and append it to the branch name until the name no longer conflicts.

It currently would work only on cname-less github pages as it needs to
specify a base directory for resources to be referenced.