Adds branch overwriting support.
Via the overwrite-branch input value, the action can simply overwrite
the branch with fresh documentation and push instead of forking the
branch and pushing the fork, so no extra step would be necessary between
the action trigger and updated docs being available in the GitHub Pages
endpoint.

Related GitHub issue: #3.